Graphics
Put the graphics banks (8-B) in bin files cause having it in text is
dumb.
Started labeling the 34 gfx files.
also realized that with different versions to compile to, the raw data
in the comment is pretty much useless cause it would be different per
version. the offsets will stay though.